+++ packages/common/src/types/ContractName.ts
@@ -80,3 +80,13 @@ export type SchemaValue = {
 export function toSchemaValue(contractName: ContractName): SchemaValue {
     return { contract: contractName.value };
 }
+
+/**
+ * Check if two contract names represent the same name of a contract.
+ * @param {ContractName} left
+ * @param {ContractName} right
+ * @returns {boolean} True if they are equal.
+ */
+export function equals(left: ContractName, right: ContractName): boolean {
+    return left.value === right.value;
+}